updating

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/expunge/ResourceExpungeService.java
@@ -23,9 +23,24 @@ package ca.uhn.fhir.jpa.dao.expunge;
 import ca.uhn.fhir.interceptor.api.HookParams;
 import ca.uhn.fhir.interceptor.api.IInterceptorBroadcaster;
 import ca.uhn.fhir.interceptor.api.Pointcut;
+import ca.uhn.fhir.jpa.api.config.DaoConfig;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
 import ca.uhn.fhir.jpa.api.dao.IFhirResourceDao;
-import ca.uhn.fhir.jpa.dao.data.*;
+import ca.uhn.fhir.jpa.dao.data.IResourceHistoryTableDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceHistoryTagDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedCompositeStringUniqueDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamCoordsDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamDateDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamNumberDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamQuantityDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamStringDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamTokenDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceIndexedSearchParamUriDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceLinkDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceProvenanceDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceTableDao;
+import ca.uhn.fhir.jpa.dao.data.IResourceTagDao;
+import ca.uhn.fhir.jpa.dao.data.ISearchParamPresentDao;
 import ca.uhn.fhir.jpa.dao.index.IdHelperService;
 import ca.uhn.fhir.jpa.model.entity.ForcedId;
 import ca.uhn.fhir.jpa.model.entity.ResourceHistoryTable;
@@ -52,7 +67,7 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Service
-class ResourceExpungeService implements IResourceExpungeService {
+public class ResourceExpungeService implements IResourceExpungeService {
 	private static final Logger ourLog = LoggerFactory.getLogger(ResourceExpungeService.class);
 
 	@Autowired
@@ -74,6 +89,8 @@ class ResourceExpungeService implements IResourceExpungeService {
 	@Autowired
 	private IResourceIndexedSearchParamNumberDao myResourceIndexedSearchParamNumberDao;
 	@Autowired
+	private IResourceIndexedCompositeStringUniqueDao myResourceIndexedCompositeStringUniqueDao;
+	@Autowired
 	private IResourceLinkDao myResourceLinkDao;
 	@Autowired
 	private IResourceTagDao myResourceTagDao;
@@ -87,6 +104,10 @@ class ResourceExpungeService implements IResourceExpungeService {
 	private DaoRegistry myDaoRegistry;
 	@Autowired
 	private IResourceProvenanceDao myResourceHistoryProvenanceTableDao;
+	@Autowired
+	private ISearchParamPresentDao mySearchParamPresentDao;
+	@Autowired
+	private DaoConfig myDaoConfig;
 
 	@Override
 	@Transactional
@@ -172,7 +193,6 @@ class ResourceExpungeService implements IResourceExpungeService {
 		theRemainingCount.addAndGet(-1 * counter.get());
 	}
 
-
 	@Override
 	@Transactional
 	public void expungeHistoricalVersionsOfIds(RequestDetails theRequestDetails, List<Long> theResourceIds, AtomicInteger theRemainingCount) {
@@ -218,24 +238,45 @@ class ResourceExpungeService implements IResourceExpungeService {
 		myResourceTableDao.deleteByPid(resource.getId());
 	}
 
-
-	@Autowired
-	private ISearchParamPresentDao mySearchParamPresentDao;
-
 	@Override
 	@Transactional
 	public void deleteAllSearchParams(Long theResourceId) {
-		myResourceIndexedSearchParamUriDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamCoordsDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamDateDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamNumberDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamQuantityDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamStringDao.deleteByResourceId(theResourceId);
-		myResourceIndexedSearchParamTokenDao.deleteByResourceId(theResourceId);
-		mySearchParamPresentDao.deleteByResourceId(theResourceId);
-		myResourceLinkDao.deleteByResourceId(theResourceId);
+		ResourceTable resource = myResourceTableDao.findById(theResourceId).orElse(null);
 
-		myResourceTagDao.deleteByResourceId(theResourceId);
+		if (resource == null || resource.isParamsUriPopulated()) {
+			myResourceIndexedSearchParamUriDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsCoordsPopulated()) {
+			myResourceIndexedSearchParamCoordsDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsDatePopulated()) {
+			myResourceIndexedSearchParamDateDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsNumberPopulated()) {
+			myResourceIndexedSearchParamNumberDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsQuantityPopulated()) {
+			myResourceIndexedSearchParamQuantityDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsStringPopulated()) {
+			myResourceIndexedSearchParamStringDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsTokenPopulated()) {
+			myResourceIndexedSearchParamTokenDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isParamsCompositeStringUniquePresent()) {
+			myResourceIndexedCompositeStringUniqueDao.deleteByResourceId(theResourceId);
+		}
+		if (myDaoConfig.getIndexMissingFields() == DaoConfig.IndexEnabledEnum.ENABLED) {
+			mySearchParamPresentDao.deleteByResourceId(theResourceId);
+		}
+		if (resource == null || resource.isHasLinks()) {
+			myResourceLinkDao.deleteByResourceId(theResourceId);
+		}
+
+		if (resource == null || resource.isHasTags()) {
+			myResourceTagDao.deleteByResourceId(theResourceId);
+		}
 	}
 
 	private void expungeHistoricalVersionsOfId(RequestDetails theRequestDetails, Long myResourceId, AtomicInteger theRemainingCount) {
